+++ lightrun.h
@@ -40,7 +40,7 @@ class Lightrun {
     uint32_t _colorsubtract(uint32_t, uint32_t);
     uint32_t _coloradd(uint32_t, uint32_t);
     uint32_t _colorscale(uint32_t color, float scale);
-    uint32_t _getColorByte(uint32_t fullcolor, int index);
+    byte _getColorByte(uint32_t fullcolor, int index);
     
     void _undoCurrentState();
     void _setNewState();
